requirements.txt
asgiref==3.4.1
dj-config-url==0.1.1
Django==3.2.16
djangorestframework==3.14.0
pip==21.3.1
psycopg2==2.9.5
pytz==2022.6
setuptools==39.0.1
sqlparse==0.4.3
typing-extensions==4.1.1
